Add ability to set ContentID on attached binaries
To be able to use an attached binary as an embedded image, you must be able to refer to it by its ContentID (e.g. <img src="cid:blah..." ...></img>.  To allow this to work, I added an optional named parameter content_id, defaulted to None, after content_type.  If set to anything other than None, the attachment's ContentId key is set to the value passed in.  What value the content id is set to is left up to the user of the API.